[feature/share]: refactor rollback dir
